Implement API `api/envs/default/apps/
implement API `api/envs/default/apps/ and refactor code
