+++ pkg/cmd/dashboard.go
@@ -1,4 +1,4 @@
-package main
+package cmd
 
 import (
 	"context"
@@ -9,15 +9,48 @@ import (
 	"syscall"
 	"time"
 
+	"sigs.k8s.io/controller-runtime/pkg/client"
+
+	ctrl "sigs.k8s.io/controller-runtime"
+
+	"github.com/cloud-native-application/rudrx/api/types"
 	"go.uber.org/zap/zapcore"
 	"gopkg.in/natefinch/lumberjack.v2"
-	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/log/zap"
 
+	cmdutil "github.com/cloud-native-application/rudrx/pkg/cmd/util"
 	"github.com/cloud-native-application/rudrx/pkg/server"
+	"github.com/spf13/cobra"
 )
 
-func main() {
+func NewDashboardCommand(c types.Args, ioStreams cmdutil.IOStreams) *cobra.Command {
+	// ctx := context.Background()
+	cmd := &cobra.Command{
+		Use:     "dashboard",
+		Short:   "Setup API Server and launch Dashboard",
+		Long:    "Setup API Server and launch Dashboard",
+		Example: `dashboard`,
+		RunE: func(cmd *cobra.Command, args []string) error {
+			newClient, err := client.New(c.Config, client.Options{Scheme: c.Schema})
+			if err != nil {
+				return err
+			}
+			SetupApiServer(newClient)
+			return nil
+		},
+		Annotations: map[string]string{
+			types.TagCommandType: types.TypeSystem,
+		},
+	}
+	cmd.SetOut(ioStreams.Out)
+	return cmd
+}
+
+/*
+SetupApiServer moved main function from cmd/server/main.go written by Ryan, so it can be integrated
+by Cli and called by cmd/server/main.go.
+*/
+func SetupApiServer(kubeClient client.Client) {
 	var logFilePath string
 	var logRetainDate int
 	var logCompress, development bool
@@ -46,7 +79,7 @@ func main() {
 
 	//Setup RESTful server
 	server := server.ApiServer{}
-	server.Launch()
+	server.Launch(kubeClient)
 	// handle signal: SIGTERM(15), SIGKILL(9)
 	sc := make(chan os.Signal, 1)
 	signal.Notify(sc, syscall.SIGTERM)
